Adding NavDrawer and other tweaks
NavDrawer only has two items in it, can add to that by adding items to
nav_draw.xml in menu.

Changing around a lot of things to use androidx, which will save some of
our hairs as we look up examples.

Added a settings fragment, so add preferences to preferences.xml, which
is currently the default fragment that shows up which will need to change.

Added some initial/placeholder icons.